refactor rename transaction details args package
