+++ wallet_presentation/src/commonMain/kotlin/net/thechance/mena/wallet/presentation/screen/transaction_details/args/TransactionDetailsArgs.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.wallet.presentation.screen.transaction_details.utils
+package net.thechance.mena.wallet.presentation.screen.transaction_details.args
 
 import kotlinx.serialization.Serializable
