+++ wallet_presentation/src/commonTest/kotlin/net/thechance/mena/wallet/presentation/screen/transaction_details/TransactionDetailsViewModelTest.kt
@@ -28,7 +28,7 @@ import net.thechance.mena.wallet.domain.model.TransactionType
 import net.thechance.mena.wallet.domain.repository.TransactionRepository
 import net.thechance.mena.wallet.presentation.base.ErrorState
 import net.thechance.mena.wallet.presentation.screen.helper.FakeStringProvider
-import net.thechance.mena.wallet.presentation.screen.transaction_details.utils.TransactionDetailsArgs
+import net.thechance.mena.wallet.presentation.screen.transaction_details.args.TransactionDetailsArgs
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
 import kotlin.test.Test
